remove float deepseekv2

+++ lmdeploy/pytorch/models/deepseek_v2.py
@@ -638,7 +638,7 @@ class DeepseekV2ForCausalLM(nn.Module):
         )
 
         logits = self.lm_head(hidden_states)
-        logits = logits.float()
+        # logits = logits.float()
         return logits
 
     def get_input_embeddings(self):
